perf: defer models config schema construction
